+++ core/src/main/scala/astraea/spark/rasterframes/util/ZeroSevenCompatibilityKit.scala
@@ -33,6 +33,10 @@ import org.apache.spark.sql.catalyst.analysis.FunctionRegistry
 import org.apache.spark.sql.functions.{lit, udf}
 import org.apache.spark.sql.rf.VersionShims
 import org.apache.spark.sql.{Column, SQLContext, TypedColumn, rf}
+import astraea.spark.rasterframes.{expressions ⇒ E, functions ⇒ F}
+
+import scala.reflect.runtime.universe._
+
 
 /**
  * UDFs for working with Tiles in Spark DataFrames.
@@ -41,215 +45,215 @@ import org.apache.spark.sql.{Column, SQLContext, TypedColumn, rf}
  */
 object ZeroSevenCompatibilityKit {
   import SparkDefaultEncoders._
-  import util._
+  //import util.NamedColumn
 
   trait RasterFunctions {
     // format: off
   /** Create a row for each cell in Tile. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def explodeTiles(cols: Column*): Column = explodeTilesSample(1.0, None, cols: _*)
 
   /** Create a row for each cell in Tile with random sampling and optional seed. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def explodeTilesSample(sampleFraction: Double, seed: Option[Long], cols: Column*): Column =
     E.ExplodeTiles(sampleFraction, seed, cols)
 
   /** Create a row for each cell in Tile with random sampling (no seed). */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def explodeTilesSample(sampleFraction: Double, cols: Column*): Column =
     E.ExplodeTiles(sampleFraction, None, cols)
 
   /** Query the number of (cols, rows) in a Tile. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileDimensions(col: Column): Column = E.GetDimensions(col)
 
   /** Flattens Tile into an array. A numeric type parameter is required. */
   @Experimental
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileToArray[T: HasCellType: TypeTag](col: Column): TypedColumn[Any, Array[T]] = withAlias("tileToArray", col)(
     udf[Array[T], Tile](F.tileToArray).apply(col)
   ).as[Array[T]]
 
   @Experimental
   /** Convert array in `arrayCol` into a Tile of dimensions `cols` and `rows`*/
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def arrayToTile(arrayCol: Column, cols: Int, rows: Int) = withAlias("array_to_tile", arrayCol)(
     udf[Tile, AnyRef](F.arrayToTile(cols, rows)).apply(arrayCol)
   )
 
   /** Create a Tile from a column of cell data with location indexes and preform cell conversion. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def assembleTile(columnIndex: Column, rowIndex: Column, cellData: Column, tileCols: Int, tileRows: Int, ct: CellType): TypedColumn[Any, Tile] =
     convertCellType(F.TileAssembler(columnIndex, rowIndex, cellData, lit(tileCols), lit(tileRows)), ct).as(cellData.columnName).as[Tile]
 
   /** Create a Tile from  a column of cell data with location indexes. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def assembleTile(columnIndex: Column, rowIndex: Column, cellData: Column, tileCols: Column, tileRows: Column): TypedColumn[Any, Tile] =
     F.TileAssembler(columnIndex, rowIndex, cellData, tileCols, tileRows)
 
   /** Extract the Tile's cell type */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def cellType(col: Column): TypedColumn[Any, CellType] = E.GetCellType(col)
 
   /** Change the Tile's cell type */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def convertCellType(col: Column, cellType: CellType): TypedColumn[Any, Tile] =
     E.SetCellType(col, cellType)
 
   /** Change the Tile's cell type */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def convertCellType(col: Column, cellTypeName: String): TypedColumn[Any, Tile] =
     E.SetCellType(col, cellTypeName)
 
   /** Convert a bounding box structure to a Geometry type. Intented to support multiple schemas. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def boundsGeometry(bounds: Column): TypedColumn[Any, Geometry] = E.BoundsToGeometry(bounds)
 
   /** Assign a `NoData` value to the Tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def withNoData(col: Column, nodata: Double) = withAlias("withNoData", col)(
     udf[Tile, Tile](F.withNoData(nodata)).apply(col)
   ).as[Tile]
 
   /**  Compute the full column aggregate floating point histogram. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def aggHistogram(col: Column): TypedColumn[Any, CellHistogram] =
   withAlias("histogram", col)(
     F.aggHistogram(col)
   ).as[CellHistogram]
 
   /** Compute the full column aggregate floating point statistics. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def aggStats(col: Column): TypedColumn[Any, CellStatistics] = withAlias("aggStats", col)(
     F.aggStats(col)
   ).as[CellStatistics]
 
   /** Computes the column aggregate mean. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def aggMean(col: Column) = CellMeanAggregate(col)
 
   /** Computes the number of non-NoData cells in a column. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def aggDataCells(col: Column) = CellCountAggregate(true, col)
 
   /** Computes the number of NoData cells in a column. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def aggNoDataCells(col: Column) = CellCountAggregate(false, col)
 
   /** Compute the Tile-wise mean */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileMean(col: Column): TypedColumn[Any, Double] =
   withAlias("tileMean", col)(
     udf[Double, Tile](F.tileMean).apply(col)
   ).as[Double]
 
   /** Compute the Tile-wise sum */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileSum(col: Column): TypedColumn[Any, Double] =
   withAlias("tileSum", col)(
     udf[Double, Tile](F.tileSum).apply(col)
   ).as[Double]
 
   /** Compute the minimum cell value in tile. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileMin(col: Column): TypedColumn[Any, Double] =
   withAlias("tileMin", col)(
     udf[Double, Tile](F.tileMin).apply(col)
   ).as[Double]
 
   /** Compute the maximum cell value in tile. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileMax(col: Column): TypedColumn[Any, Double] =
   withAlias("tileMax", col)(
     udf[Double, Tile](F.tileMax).apply(col)
   ).as[Double]
 
   /** Compute TileHistogram of Tile values. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileHistogram(col: Column): TypedColumn[Any, CellHistogram] =
   withAlias("tileHistogram", col)(
     udf[CellHistogram, Tile](F.tileHistogram).apply(col)
   ).as[CellHistogram]
 
   /** Compute statistics of Tile values. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileStats(col: Column): TypedColumn[Any, CellStatistics] =
   withAlias("tileStats", col)(
     udf[CellStatistics, Tile](F.tileStats).apply(col)
   ).as[CellStatistics]
 
   /** Counts the number of non-NoData cells per Tile. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def dataCells(tile: Column): TypedColumn[Any, Long] =
     withAlias("dataCells", tile)(
       udf(F.dataCells).apply(tile)
     ).as[Long]
 
   /** Counts the number of NoData cells per Tile. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def noDataCells(tile: Column): TypedColumn[Any, Long] =
     withAlias("noDataCells", tile)(
       udf(F.noDataCells).apply(tile)
     ).as[Long]
 
 
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def isNoDataTile(tile: Column): TypedColumn[Any, Boolean] =
     withAlias("isNoDataTile", tile)(
       udf(F.isNoDataTile).apply(tile)
     ).as[Boolean]
 
   /** Compute cell-local aggregate descriptive statistics for a column of Tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localAggStats(col: Column): Column =
   withAlias("localAggStats", col)(
     F.localAggStats(col)
   )
 
   /** Compute the cell-wise/local max operation between Tiles in a column. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localAggMax(col: Column): TypedColumn[Any, Tile] =
   withAlias("localAggMax", col)(
     F.localAggMax(col)
   ).as[Tile]
 
   /** Compute the cellwise/local min operation between Tiles in a column. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localAggMin(col: Column): TypedColumn[Any, Tile] =
   withAlias("localAggMin", col)(
     F.localAggMin(col)
   ).as[Tile]
 
   /** Compute the cellwise/local mean operation between Tiles in a column. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localAggMean(col: Column): TypedColumn[Any, Tile] =
   withAlias("localAggMean", col)(
     F.localAggMean(col)
   ).as[Tile]
 
   /** Compute the cellwise/local count of non-NoData cells for all Tiles in a column. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localAggDataCells(col: Column): TypedColumn[Any, Tile] =
   withAlias("localCount", col)(
     F.localAggCount(col)
   ).as[Tile]
 
   /** Compute the cellwise/local count of NoData cells for all Tiles in a column. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localAggNoDataCells(col: Column): TypedColumn[Any, Tile] =
   withAlias("localNodataCount", col)(
     F.localAggNodataCount(col)
   ).as[Tile]
 
   /** Cellwise addition between two Tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localAdd(left: Column, right: Column): TypedColumn[Any, Tile] =
   withAlias("local_add", left, right)(
     udf(F.localAdd).apply(left, right)
   ).as[Tile]
 
   /** Cellwise addition of a scalar to a tile. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localAddScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match {
       case i: Int => F.localAddScalarInt(_: Tile, i)
@@ -260,14 +264,14 @@ object ZeroSevenCompatibilityKit {
   }
 
   /** Cellwise subtraction between two Tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localSubtract(left: Column, right: Column): TypedColumn[Any, Tile] =
   withAlias("localSubtract", left, right)(
     udf(F.localSubtract).apply(left, right)
   ).as[Tile]
 
   /** Cellwise subtraction of a scalar from a tile. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localSubtractScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match {
       case i: Int => F.localSubtractScalarInt(_: Tile, i)
@@ -278,14 +282,14 @@ object ZeroSevenCompatibilityKit {
   }
 
   /** Cellwise multiplication between two Tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localMultiply(left: Column, right: Column): TypedColumn[Any, Tile] =
   withAlias("localMultiply", left, right)(
     udf(F.localMultiply).apply(left, right)
   ).as[Tile]
 
   /** Cellwise multiplication of a tile by a scalar. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localMultiplyScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match {
       case i: Int => F.localMultiplyScalarInt(_: Tile, i)
@@ -296,14 +300,14 @@ object ZeroSevenCompatibilityKit {
   }
 
   /** Cellwise division between two Tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localDivide(left: Column, right: Column): TypedColumn[Any, Tile] =
   withAlias("localDivide", left, right)(
     udf(F.localDivide).apply(left, right)
   ).as[Tile]
 
   /** Cellwise division of a tile by a scalar. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localDivideScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match {
       case i: Int => F.localDivideScalarInt(_: Tile, i)
@@ -314,7 +318,7 @@ object ZeroSevenCompatibilityKit {
   }
 
   /** Perform an arbitrary GeoTrellis `LocalTileBinaryOp` between two Tile columns. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localAlgebra(op: LocalTileBinaryOp, left: Column, right: Column):
   TypedColumn[Any, Tile] =
     withAlias(opName(op), left, right)(
@@ -322,43 +326,43 @@ object ZeroSevenCompatibilityKit {
     ).as[Tile]
 
   /** Compute the normalized difference of two tile columns */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def normalizedDifference(left: Column, right: Column): TypedColumn[Any, Tile] =
     withAlias("normalizedDifference", left, right)(
       udf(F.normalizedDifference).apply(left, right)
     ).as[Tile]
 
   /** Constructor for constant tile column */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def makeConstantTile(value: Number, cols: Int, rows: Int, cellType: String): TypedColumn[Any, Tile] =
     udf(() => F.makeConstantTile(value, cols, rows, cellType)).apply().as(s"constant_$cellType").as[Tile]
 
   /** Alias for column of constant tiles of zero */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileZeros(cols: Int, rows: Int, cellType: String = "float64"): TypedColumn[Any, Tile] =
     udf(() => F.tileZeros(cols, rows, cellType)).apply().as(s"zeros_$cellType").as[Tile]
 
   /** Alias for column of constant tiles of one */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def tileOnes(cols: Int, rows: Int, cellType: String = "float64"): TypedColumn[Any, Tile] =
     udf(() => F.tileOnes(cols, rows, cellType)).apply().as(s"ones_$cellType").as[Tile]
 
   /** Where the mask tile equals the mask value, replace values in the source tile with NODATA */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def maskByValue(sourceTile: Column, maskTile: Column, maskValue: Column): TypedColumn[Any, Tile] =
     withAlias("maskByValue", sourceTile, maskTile, maskValue)(
       udf(F.maskByValue).apply(sourceTile, maskTile, maskValue)
     ).as[Tile]
 
   /** Where the mask tile DOES NOT contain NODATA, replace values in the source tile with NODATA */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def inverseMask(sourceTile: Column, maskTile: Column): TypedColumn[Any, Tile] =
     withAlias("inverseMask", sourceTile, maskTile)(
       udf(F.inverseMask).apply(sourceTile, maskTile)
     ).as[Tile]
 
   /** Reproject a column of geometry from one CRS to another. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def reprojectGeometry(sourceGeom: Column, srcCRS: CRS, dstCRS: CRS): TypedColumn[Any, Geometry] =
     withAlias("reprojectGeometry", sourceGeom)(
       udf(F.reprojectGeometry(_: Geometry, srcCRS, dstCRS)).apply(sourceGeom)
@@ -366,14 +370,14 @@ object ZeroSevenCompatibilityKit {
 
   /** Render Tile as ASCII string for debugging purposes. */
   @Experimental
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def renderAscii(col: Column): TypedColumn[Any, String] =
   withAlias("renderAscii", col)(
     udf[String, Tile](F.renderAscii).apply(col)
   ).as[String]
 
   /** Cellwise less than value comparison between two tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localLess(left: Column, right: Column): TypedColumn[Any, Tile] =
     withAlias("localLess", left, right)(
       udf(F.localLess).apply(left, right)
@@ -381,7 +385,7 @@ object ZeroSevenCompatibilityKit {
 
 
   /** Cellwise less than value comparison between a tile and a scalar. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localLessScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match{
       case i: Int => F.localLessScalarInt(_: Tile, i)
@@ -391,14 +395,14 @@ object ZeroSevenCompatibilityKit {
   }
 
   /** Cellwise less than or equal to value comparison between a tile and a scalar. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localLessEqual(left: Column, right: Column): TypedColumn[Any, Tile]  =
     withAlias("localLessEqual", left, right)(
       udf(F.localLess).apply(left, right)
     ).as[Tile]
 
   /** Cellwise less than or equal to value comparison between a tile and a scalar. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localLessEqualScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match{
       case i: Int => F.localLessEqualScalarInt(_: Tile, i)
@@ -408,7 +412,7 @@ object ZeroSevenCompatibilityKit {
   }
 
   /** Cellwise greater than value comparison between two tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localGreater(left: Column, right: Column): TypedColumn[Any, Tile] =
     withAlias("localGreater", left, right)(
       udf(F.localGreater).apply(left, right)
@@ -416,7 +420,7 @@ object ZeroSevenCompatibilityKit {
 
 
   /** Cellwise greater than value comparison between a tile and a scalar. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localGreaterScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match{
       case i: Int => F.localGreaterScalarInt(_: Tile, i)
@@ -426,14 +430,14 @@ object ZeroSevenCompatibilityKit {
   }
 
   /** Cellwise greater than or equal to value comparison between two tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localGreaterEqual(left: Column, right: Column): TypedColumn[Any, Tile]  =
     withAlias("localGreaterEqual", left, right)(
       udf(F.localGreaterEqual).apply(left, right)
     ).as[Tile]
 
   /** Cellwise greater than or equal to value comparison between a tile and a scalar. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localGreaterEqualScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match{
       case i: Int => F.localGreaterEqualScalarInt(_: Tile, i)
@@ -443,14 +447,14 @@ object ZeroSevenCompatibilityKit {
   }
 
   /** Cellwise equal to value comparison between two tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localEqual(left: Column, right: Column): TypedColumn[Any, Tile]  =
     withAlias("localEqual", left, right)(
       udf(F.localEqual).apply(left, right)
     ).as[Tile]
 
   /** Cellwise equal to value comparison between a tile and a scalar. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localEqualScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match{
       case i: Int => F.localEqualScalarInt(_: Tile, i)
@@ -459,14 +463,14 @@ object ZeroSevenCompatibilityKit {
     udf(f).apply(tileCol).as(s"localEqualScalar($tileCol, $value)").as[Tile]
   }
   /** Cellwise inequality comparison between two tiles. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localUnequal(left: Column, right: Column): TypedColumn[Any, Tile]  =
     withAlias("localUnequal", left, right)(
       udf(F.localUnequal).apply(left, right)
     ).as[Tile]
 
   /** Cellwise inequality comparison between a tile and a scalar. */
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x. Please use \"snake_case\" variant instead.", "0.8.0")
   def localUnequalScalar[T: Numeric](tileCol: Column, value: T): TypedColumn[Any, Tile] = {
     val f = value match{
       case i: Int => F.localUnequalScalarInt(_: Tile, i)
@@ -475,7 +479,7 @@ object ZeroSevenCompatibilityKit {
     udf(f).apply(tileCol).as(s"localUnequalScalar($tileCol, $value)").as[Tile]
   }
 }
-  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x", "0.8.0")
+  @deprecated("Part of 0.7.x compatility kit, to be removed after 0.8.x.", "0.8.0")
   def register(sqlContext: SQLContext): Unit = {
 
     /** Unary expression builder builder. */