+++ core/src/main/scala/astraea/spark/rasterframes/util/package.scala
@@ -36,10 +36,8 @@ import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.rf._
 import org.apache.spark.sql.types.StringType
 import org.apache.spark.sql.{Column, DataFrame, SQLContext}
-import org.osgeo.proj4j.Proj4jException
 
 import scala.Boolean.box
-import scala.util.control.NonFatal
 
 /**
  * Internal utilities.