+++ pyrasterframes/python/examples/__init__.py
@@ -14,6 +14,9 @@ if len(jarpath) > 0:
         .config('spark.driver.extraClassPath', pyJar)
         .config('spark.executor.extraClassPath', pyJar)
         .config("spark.ui.enabled", "false")
+        .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
+        .config("spark.kryo.registrator", "astraea.spark.rasterframes.util.RFKryoRegistrator")
+        .config("spark.kryoserializer.buffer.max", "500m")
         .getOrCreate())