+++ pyrasterframes/python/tests/PyRasterFramesTests.py
@@ -36,6 +36,9 @@ class RasterFunctionsTest(unittest.TestCase):
         cls.spark = (SparkSession.builder
             .config('spark.driver.extraClassPath', jarpath)
             .config('spark.executor.extraClassPath', jarpath)
+            .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
+            .config("spark.kryo.registrator", "astraea.spark.rasterframes.util.RFKryoRegistrator")
+            .config("spark.kryoserializer.buffer.max", "500m")
             .getOrCreate())
         cls.spark.sparkContext.setLogLevel('ERROR')
         print(cls.spark.version)